refactor: extract ref watcher from API
